src/requirements.txt
# For python 2.7: install in lib/ + use appengine_config.py
# $ pip install -t lib -r requirements.txt
# For python 3.7: simply have this file ready for deployment
wsgidav>=3.0.1
future>=0.18.2
Flask>=1.1.1
cachelib>=0.1
#google-auth>=1.6.3
#requests>=2.22.0
#CacheControl>=0.12.5
google-cloud-datastore>=1.10.0
#google-cloud-firestore>=1.6.0
#firebase-admin>=3.2.0
